# Projeto/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "APP - Hotelaria - Sistema desenvolvido ao longo da discipilina de tecWeb - Ufsj ",
    "version": "1.0.0",
    "title": "ClickHotel",
    "contact": {
      "email": "anaflavia.lemos1@gmail.com"
    }
  },
  "host": "clickbus.ccomp.ufsj.edu.br",
  "tags": [
    {
      "name": "Hoteis",
      "description": "Descrição dos hoteis"
    },
    {
      "name": "Reservas",
      "description": "Tudo sobre as reservas"
    },
    {
      "name": "usuario",
      "description": "Operações para gerência de usuários"
    }
  ],
  "schemes": [
    "http"
  ],
  "paths": {
    "/hoteis": {
      "post": {
        "tags": [
          "hoteis"
        ],
        "summary": "Adiciona uma novo hotel ao sistema ClickHotel",
        "operationId": "addHoteis",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Objeto Hotel a ser adicionado no clickhotel",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Hoteis"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Hoteis"
            }
          }
        }
      }
    },
    "/hoteis/{hoteisId}": {
      "get": {
        "tags": [
          "hoteis"
        ],
        "summary": "Busca os hoteis pelo ID",
        "operationId": "getHoteisById",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "hoteisId",
            "in": "path",
            "description": "ID of hoteis to return",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfull operation",
            "schema": {
              "$ref": "#/definitions/Hoteis"
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Hoteis not found"
          }
        }
      },
      "put": {
        "tags": [
          "hoteis"
        ],
        "summary": "Atualizar uma hotel no sistema",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "hoteisId",
            "in": "path",
            "description": "ID that need to be updated",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "body",
            "in": "body",
            "description": "updated hotel object",
            "schema": {
              "$ref": "#/definitions/Hoteis"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfull operation",
            "schema": {
              "$ref": "#/definitions/Hoteis"
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Hoteis not found"
          }
        }
      },
      "delete": {
        "tags": [
          "hoteis"
        ],
        "summary": "Apaga um hotel",
        "operationId": "deletaHoteis",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "hoteisId",
            "in": "path",
            "description": "Hotel ID to delete",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfull operation",
            "schema": {
              "$ref": "#/definitions/Hoteis"
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Hoteis not found"
          }
        }
      }
    }
  },
  "definitions": {
    "Hoteis": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "nome_hoteis": {
          "type": "string"
        },
        "dias_reserva": {
          "type": "integer",
          "format": "int64"
        },
        "modelo_hotel": {
          "type": "string"
        },
        "status_hotel": {
          "type": "string",
          "enum": [
            "available",
            "rented",
            "returned"
          ]
        }
      }
    }
  }
}